Add files via upload
As a project work of a Electronic Shop Practice course, we designed a Soccer robot. Basically, this soccer robot does not refer to a robot that has feet and hands but rather a car automated by servo motors which were programmed to kick a ball. There, we used a Arduino uno board, Bluetooth module, servo motors, motor driver etc. There is given a short video of our project as a proof of accomplishment of our project successfully.